+++ src/createGridComponent.js
@@ -2,6 +2,9 @@
 
 import memoizeOne from 'memoize-one';
 import { createElement, PureComponent } from 'react';
+import { cancelTimeout, requestTimeout } from './timer';
+
+import type { TimeoutID } from './timer';
 
 export type ScrollToAlign = 'auto' | 'center' | 'start' | 'end';
 
@@ -50,7 +53,8 @@ export type Props<T> = {|
   initialScrollLeft?: number,
   initialScrollTop?: number,
   innerRef?: any,
-  innerTagName?: string,
+  innerElementType?: React$ElementType,
+  innerTagName?: string, // deprecated
   itemData: T,
   itemKey?: (params: {|
     columnIndex: number,
@@ -60,8 +64,11 @@ export type Props<T> = {|
   onItemsRendered?: OnItemsRenderedCallback,
   onScroll?: OnScrollCallback,
   outerRef?: any,
-  outerTagName?: string,
-  overscanCount: number,
+  outerElementType?: React$ElementType,
+  outerTagName?: string, // deprecated
+  overscanColumnsCount?: number,
+  overscanCount?: number, // deprecated
+  overscanRowsCount?: number,
   rowCount: number,
   rowHeight: itemSize,
   style?: Object,
@@ -154,10 +161,7 @@ export default function createGridComponent({
     _outerRef: ?HTMLDivElement;
 
     static defaultProps = {
-      innerTagName: 'div',
       itemData: undefined,
-      outerTagName: 'div',
-      overscanCount: 1,
       useIsScrolling: false,
     };
 
@@ -272,7 +276,7 @@ export default function createGridComponent({
 
     componentWillUnmount() {
       if (this._resetIsScrollingTimeoutId !== null) {
-        clearTimeout(this._resetIsScrollingTimeoutId);
+        cancelTimeout(this._resetIsScrollingTimeoutId);
       }
     }
 
@@ -283,9 +287,11 @@ export default function createGridComponent({
         columnCount,
         height,
         innerRef,
+        innerElementType,
         innerTagName,
         itemData,
         itemKey = defaultItemKey,
+        outerElementType,
         outerTagName,
         rowCount,
         style,
@@ -338,7 +344,7 @@ export default function createGridComponent({
       );
 
       return createElement(
-        ((outerTagName: any): string),
+        outerElementType || outerTagName || 'div',
         {
           className,
           onScroll: this._onScroll,
@@ -353,7 +359,7 @@ export default function createGridComponent({
             ...style,
           },
         },
-        createElement(((innerTagName: any): string), {
+        createElement(innerElementType || innerTagName || 'div', {
           children: items,
           ref: innerRef,
           style: {
@@ -503,8 +509,16 @@ export default function createGridComponent({
     _getItemStyleCache = memoizeOne((_: any, __: any) => ({}));
 
     _getHorizontalRangeToRender(): [number, number, number, number] {
-      const { columnCount, overscanCount, rowCount } = this.props;
-      const { horizontalScrollDirection, scrollLeft } = this.state;
+      const {
+        columnCount,
+        overscanColumnsCount,
+        overscanCount,
+        rowCount,
+      } = this.props;
+      const { horizontalScrollDirection, isScrolling, scrollLeft } = this.state;
+
+      const overscanCountResolved: number =
+        overscanColumnsCount || overscanCount || 1;
 
       if (columnCount === 0 || rowCount === 0) {
         return [0, 0, 0, 0];
@@ -525,12 +539,12 @@ export default function createGridComponent({
       // Overscan by one item in each direction so that tab/focus works.
       // If there isn't at least one extra item, tab loops back around.
       const overscanBackward =
-        horizontalScrollDirection === 'backward'
-          ? Math.max(1, overscanCount)
+        !isScrolling || horizontalScrollDirection === 'backward'
+          ? Math.max(1, overscanCountResolved)
           : 1;
       const overscanForward =
-        horizontalScrollDirection === 'forward'
-          ? Math.max(1, overscanCount)
+        !isScrolling || horizontalScrollDirection === 'forward'
+          ? Math.max(1, overscanCountResolved)
           : 1;
 
       return [
@@ -542,8 +556,16 @@ export default function createGridComponent({
     }
 
     _getVerticalRangeToRender(): [number, number, number, number] {
-      const { columnCount, rowCount, overscanCount } = this.props;
-      const { verticalScrollDirection, scrollTop } = this.state;
+      const {
+        columnCount,
+        overscanCount,
+        overscanRowsCount,
+        rowCount,
+      } = this.props;
+      const { isScrolling, verticalScrollDirection, scrollTop } = this.state;
+
+      const overscanCountResolved: number =
+        overscanRowsCount || overscanCount || 1;
 
       if (columnCount === 0 || rowCount === 0) {
         return [0, 0, 0, 0];
@@ -564,9 +586,13 @@ export default function createGridComponent({
       // Overscan by one item in each direction so that tab/focus works.
       // If there isn't at least one extra item, tab loops back around.
       const overscanBackward =
-        verticalScrollDirection === 'backward' ? Math.max(1, overscanCount) : 1;
+        !isScrolling || verticalScrollDirection === 'backward'
+          ? Math.max(1, overscanCountResolved)
+          : 1;
       const overscanForward =
-        verticalScrollDirection === 'forward' ? Math.max(1, overscanCount) : 1;
+        !isScrolling || verticalScrollDirection === 'forward'
+          ? Math.max(1, overscanCountResolved)
+          : 1;
 
       return [
         Math.max(0, startIndex - overscanBackward),
@@ -620,21 +646,10 @@ export default function createGridComponent({
 
     _resetIsScrollingDebounced = () => {
       if (this._resetIsScrollingTimeoutId !== null) {
-        clearTimeout(this._resetIsScrollingTimeoutId);
+        cancelTimeout(this._resetIsScrollingTimeoutId);
       }
 
-      this._resetIsScrollingTimeoutId = setTimeout(
-        this._resetIsScrolling,
-        IS_SCROLLING_DEBOUNCE_INTERVAL
-      );
-    };
-
-    _resetIsScrollingDebounced = () => {
-      if (this._resetIsScrollingTimeoutId !== null) {
-        clearTimeout(this._resetIsScrollingTimeoutId);
-      }
-
-      this._resetIsScrollingTimeoutId = setTimeout(
+      this._resetIsScrollingTimeoutId = requestTimeout(
         this._resetIsScrolling,
         IS_SCROLLING_DEBOUNCE_INTERVAL
       );
@@ -652,8 +667,29 @@ export default function createGridComponent({
   };
 }
 
-const validateSharedProps = ({ children, height, width }: Props<any>): void => {
+const validateSharedProps = ({
+  children,
+  height,
+  innerTagName,
+  outerTagName,
+  overscanCount,
+  width,
+}: Props<any>): void => {
   if (process.env.NODE_ENV !== 'production') {
+    if (typeof overscanCount === 'number') {
+      console.warn(
+        'The overscanCount prop has been deprecated. ' +
+          'Please use the overscanColumnsCount and overscanRowsCount props instead.'
+      );
+    }
+
+    if (innerTagName != null || outerTagName != null) {
+      console.warn(
+        'The innerTagName and outerTagName props have been deprecated. ' +
+          'Please use the innerElementType and outerElementType props instead.'
+      );
+    }
+
     if (children == null) {
       throw Error(
         'An invalid "children" prop has been specified. ' +