+++ src/createListComponent.js
@@ -2,6 +2,9 @@
 
 import memoizeOne from 'memoize-one';
 import { createElement, PureComponent } from 'react';
+import { cancelTimeout, requestTimeout } from './timer';
+
+import type { TimeoutID } from './timer';
 
 export type ScrollToAlign = 'auto' | 'center' | 'start' | 'end';
 
@@ -40,7 +43,8 @@ export type Props<T> = {|
   height: number | string,
   initialScrollOffset?: number,
   innerRef?: any,
-  innerTagName?: string,
+  innerElementType?: React$ElementType,
+  innerTagName?: string, // deprecated
   itemCount: number,
   itemData: T,
   itemKey?: (index: number, data: T) => any,
@@ -48,7 +52,8 @@ export type Props<T> = {|
   onItemsRendered?: onItemsRenderedCallback,
   onScroll?: onScrollCallback,
   outerRef?: any,
-  outerTagName?: string,
+  outerElementType?: React$ElementType,
+  outerTagName?: string, // deprecated
   overscanCount: number,
   style?: Object,
   useIsScrolling: boolean,
@@ -126,9 +131,7 @@ export default function createListComponent({
 
     static defaultProps = {
       direction: 'vertical',
-      innerTagName: 'div',
       itemData: undefined,
-      outerTagName: 'div',
       overscanCount: 2,
       useIsScrolling: false,
     };
@@ -217,7 +220,7 @@ export default function createListComponent({
 
     componentWillUnmount() {
       if (this._resetIsScrollingTimeoutId !== null) {
-        clearTimeout(this._resetIsScrollingTimeoutId);
+        cancelTimeout(this._resetIsScrollingTimeoutId);
       }
     }
 
@@ -228,10 +231,12 @@ export default function createListComponent({
         direction,
         height,
         innerRef,
+        innerElementType,
         innerTagName,
         itemCount,
         itemData,
         itemKey = defaultItemKey,
+        outerElementType,
         outerTagName,
         style,
         useIsScrolling,
@@ -269,7 +274,7 @@ export default function createListComponent({
       );
 
       return createElement(
-        ((outerTagName: any): string),
+        outerElementType || outerTagName || 'div',
         {
           className,
           onScroll,
@@ -284,7 +289,7 @@ export default function createListComponent({
             ...style,
           },
         },
-        createElement(((innerTagName: any): string), {
+        createElement(innerElementType || innerTagName || 'div', {
           children: items,
           ref: innerRef,
           style: {
@@ -377,7 +382,8 @@ export default function createListComponent({
       const { direction, itemSize } = this.props;
 
       const itemStyleCache = this._getItemStyleCache(
-        shouldResetStyleCacheOnItemSizeChange && itemSize
+        shouldResetStyleCacheOnItemSizeChange && itemSize,
+        shouldResetStyleCacheOnItemSizeChange && direction
       );
 
       let style;
@@ -408,12 +414,12 @@ export default function createListComponent({
       return style;
     };
 
-    _getItemStyleCache: (_: any) => ItemStyleCache;
-    _getItemStyleCache = memoizeOne((_: any) => ({}));
+    _getItemStyleCache: (_: any, __: any) => ItemStyleCache;
+    _getItemStyleCache = memoizeOne((_: any, __: any) => ({}));
 
     _getRangeToRender(): [number, number, number, number] {
       const { itemCount, overscanCount } = this.props;
-      const { scrollDirection, scrollOffset } = this.state;
+      const { isScrolling, scrollDirection, scrollOffset } = this.state;
 
       if (itemCount === 0) {
         return [0, 0, 0, 0];
@@ -434,9 +440,13 @@ export default function createListComponent({
       // Overscan by one item in each direction so that tab/focus works.
       // If there isn't at least one extra item, tab loops back around.
       const overscanBackward =
-        scrollDirection === 'backward' ? Math.max(1, overscanCount) : 1;
+        !isScrolling || scrollDirection === 'backward'
+          ? Math.max(1, overscanCount)
+          : 1;
       const overscanForward =
-        scrollDirection === 'forward' ? Math.max(1, overscanCount) : 1;
+        !isScrolling || scrollDirection === 'forward'
+          ? Math.max(1, overscanCount)
+          : 1;
 
       return [
         Math.max(0, startIndex - overscanBackward),
@@ -504,10 +514,10 @@ export default function createListComponent({
 
     _resetIsScrollingDebounced = () => {
       if (this._resetIsScrollingTimeoutId !== null) {
-        clearTimeout(this._resetIsScrollingTimeoutId);
+        cancelTimeout(this._resetIsScrollingTimeoutId);
       }
 
-      this._resetIsScrollingTimeoutId = setTimeout(
+      this._resetIsScrollingTimeoutId = requestTimeout(
         this._resetIsScrolling,
         IS_SCROLLING_DEBOUNCE_INTERVAL
       );
@@ -519,7 +529,7 @@ export default function createListComponent({
       this.setState({ isScrolling: false }, () => {
         // Clear style cache after state update has been committed.
         // This way we don't break pure sCU for items that don't use isScrolling param.
-        this._getItemStyleCache(-1);
+        this._getItemStyleCache(-1, null);
       });
     };
   };
@@ -535,9 +545,18 @@ const validateSharedProps = ({
   children,
   direction,
   height,
+  innerTagName,
+  outerTagName,
   width,
 }: Props<any>): void => {
   if (process.env.NODE_ENV !== 'production') {
+    if (innerTagName != null || outerTagName != null) {
+      console.warn(
+        'The innerTagName and outerTagName props have been deprecated. ' +
+          'Please use the innerElementType and outerElementType props instead.'
+      );
+    }
+
     if (direction !== 'horizontal' && direction !== 'vertical') {
       throw Error(
         'An invalid "direction" prop has been specified. ' +